+++ explosion.h
@@ -11,6 +11,6 @@ extern void explosion_spark_reset(Entity* spark, Vec2 *position, uint32_t life_t
 
 extern void explosion_spark_fixed_update(Entity* spark);
 
-extern void explosion_spawn(uint8_t num_sparks, Vec2 *position, uint32_t life_time);
+extern void explosion_spawn(uint32_t num_sparks, Vec2 *position, uint32_t life_time);
 
 #endif /* _EXPLOSION_H */
